requirements.txt
fastapi==0.115.2
uvicorn==0.30.1
httpx==0.27.0
numpy==1.26.4
pydantic==2.8.2
pyttsx3==2.90
psycopg[binary]==3.1.19
python-dotenv==1.0.1

